Fix user lock update detection logic (#10436)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
`UsersController` was incorrectly detecting if an update had been
requested to a user's locked status by checking the presence of the form
value, but that value was always present in the form. This prevented
other user updates from being made by an admin.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
It now just checks if the form value for locked is different from the
current locked status of the user.



<!-- If there are any visual changes, please attach images, videos, or
gifs. -->