Fix development packages (#10437)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
Got an error in development when `convert` wasn't found - `imagemagick`
was not included in the Dockerfile for the devcontainer. Also, I noticed
`wkhtmltopdf` was included in the packages you need to install if
setting up HCB locally, but it seems that binary is provided by the
`wkhtmltopdf-binary` in the `Gemfile`.


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Added `imagemagick` to the Dockerfile and removed the `wkhtmltopdf`
package from `development.md`


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->