Fix event tag creation (#10435)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
Event tag creation seemed to be a little broken - when you created one
from the admin tab of an event's settings, it would create the tag but
wouldn't link it due to the slug that rails sets as `params[:event_id]`
being used as a numeric record ID.


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Removed the redundant hidden field from the new event tag form and
changed the controller to interpret the ID as the event's slug.


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->