Use wkhtmltopdf-binary to ensure consistent version (#10472)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
Currently, `wkhtmltopdf` is being installed as an apt package on
production but using the `wkhtmltopdf-binary` gem on development. There
could be a version mismatch between production and development, which
would make a bug related to this package confusing to solve.


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Removes installing `wkhtmltopdf` on production and always use the
`wkhtmltopdf-binary` gem instead.


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->