Fix error in control policy (#10494)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
There is currently an error raised when a user not in an event tries to
access a spending control page, when it should just display the normal
"You are not authorized to perform this action." toast.


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Added a `&` to the policy method that finds the organizer position in
case it does not exist.


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->