Fix memo underline after stream (#10481)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
Underlines on linked memos were disappearing after a popover was opened.
The `memo_stream` partial was rendering a turbo stream that replaced the
memo with a memo without the underline.


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Moved the underline style to the parent of the span that is replaced by
`memo_stream`.


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->