Enable updating tags from HCB code popover (#10482)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
Tags cannot currently be updated in the HCB code popover, even though
they display there.


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Updated menus to have a z-index above that of modals so that the tag
menu shows and removed `z-index: 999` styles from other menus already
displaying on top of popovers. Also removed the checks that disabled
updating tags from the popover and fixed the streams to update in all
locations where tags are rendered.

![image](https://github.com/user-attachments/assets/1bb0da6d-1ce5-430e-b19a-e9498c698c41)




<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

---------

Co-authored-by: Gary Tou <gary@garytou.com>