Unpin IRB version (#10701)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
The IRB version was pinned to 1.14.3 since console1984 did not work with
the latest version. This has been fixed in the latest version of
console1984.


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Unpinned IRB in the Gemfile. Tested in development with console1984
enabled to make sure it works.


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->