[Receipt Bin] Fix `per` param (#10705)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->

We use the `per` param to determine the layout of the page (whether
transactions are rendered in groups, or a single list). Unfortunately,
it 500s at the moment because we're comparing an integer to a string
(`params[:per` is a string).

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->

This fixes it by casting `per` to an int.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->