[Transactions] Improve consistency in views (#10157)
Before, the headings for transactions did not follow a consistent
structure and also had differences in their text formatting. This PR
aims to address those issues and make the headings more consistent and
also set a standard for text style (names get none, background text is
"regular muted", and money is "regular").

---------

Co-authored-by: Gary Tou <gary@garytou.com>
Co-authored-by: David Cornu <me@davidcornu.com>