Small fixes for one time use grants (#10700)
## Summary of the problem
One time use grants weren't setting a papertrail when frozen. The
location in the details was also bad and the grantee had no indication
of it being one time use.


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Adds papertrail to be bank@hackclub.com when frozen, moves the detail
row, and adds copy to show one time use to the grantee.


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->