Hide last4 of card number in transparent mode (#10773)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
Showing the last 4 digits of credit cards in transparency mode could be
dangerous!


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Removes showing the last 4 digits of a card number when a member of the
event is not logged in - readers and up can still see last 4, but only
the cardholder can see the full number.


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->