Add email for when a donation goal is reached (#10723)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
Organizers do not currently get any notification when their donation
goal has been reached.


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Added an email that is sent to all organizers when the goal has been
reached, along with a summary of the donations that contributed to it
and a link to adjust the goal.


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

---------

Co-authored-by: Gary Tou <gary@garytou.com>