Theme switching fixes (#10789)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
https://github.com/hackclub/hcb/pull/10757 introduced a couple bugs with
theme switching where the selected theme wouldn't get a checkmark on
page load and users were unable to switch on an admin page.


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Fixed the bugs by falling back to the system theme when there was no
theme cookie and adding the full blog widget to the admin header


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->