Fix blog unread indicator (#10786)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
You can't fetch unreads from production in development due to CORS, and
it's useful to have any development changes as well. The iframe was also
loading every time the page loaded, even when the user did not open the
widget, causing the indicator to reset

Closes #10777


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Expects blog to be on `localhost:3001` and fetches unreads from that
host when in development. Enforced color and added lazy loading to
iframe


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

![image](https://github.com/user-attachments/assets/2d8642cd-7537-4ef9-8b1c-2f22be33c7ab)