[Org Home] Disable transactions if organizers aren't signed in (#10647)
This pull request refines the logic for rendering transaction links in
the `app/views/events/home` directory. The changes ensure that links are
only clickable when the user is signed in as an organizer, improving the
user experience and preventing unintended behavior for non-organizers.

### Updates to transaction link rendering:

*
[`app/views/events/home/_money_movement.html.erb`](diffhunk://#diff-9d88ddf84e51d3d3103283019be589083a086d5f5a5d427078773bd62a62fbb4L13-R20):
Updated the "money in" and "money out" transaction links to include a
conditional check for `organizer_signed_in?`. Links are now only
rendered with `href` and associated attributes if the user is an
organizer.
[[1]](diffhunk://#diff-9d88ddf84e51d3d3103283019be589083a086d5f5a5d427078773bd62a62fbb4L13-R20)
[[2]](diffhunk://#diff-9d88ddf84e51d3d3103283019be589083a086d5f5a5d427078773bd62a62fbb4L51-R65)

*
[`app/views/events/home/_transactions.html.erb`](diffhunk://#diff-f4b1ba8fc2c80c69b587c22ee12857e311dc00dcaee1208323ad47d27b8200f9L5-R13):
Applied the same `organizer_signed_in?` conditional logic to the "recent
transactions" links, ensuring consistent behavior across transaction
displays.