[Admin] Add risk level dropdown at org creation (#10861)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
There is no easy way for operations to determine the risk level of an
organization when creating it.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Created a dropdown at the bottom of admin org creation page that can be
used to set the organization risk level

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

![image](https://github.com/user-attachments/assets/e9d6d799-3341-4ea9-ad05-33144d1f45de)