Fix announcement publishing emails (#10879)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
The check in `AnnouncementPublishedJob` to make sure announcements are
still published when the job runs was never sending emails since it had
an old version of the record


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Updated the announcement record before passing it to the job


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->