Fix overflows in user and event navs (#10880)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
The "More" menu was being cut off on the mobile event nav with the
announcements button added, and Settings was being cut off on the mobile
user nav.


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Moved announcements and feed to the More menu, and hide feed when no
events are being followed.
![Screen Shot 2025-07-09 at 11 26
46](https://github.com/user-attachments/assets/5218d705-7afa-48c8-b55f-15544a58b051)
![Screen Shot 2025-07-09 at 11 42
40](https://github.com/user-attachments/assets/3f18c9c1-9d00-4250-af0b-2f1c986a3e01)




<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

---------

Co-authored-by: David Cornu <davidc@hackclub.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>