Use AASM for announcement draft state (#10946)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
Preparing for announcement prompting which will need the state to be
AASM


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Added AASM state for announcement draft state and used its helpers.
Added script to backfill the column which should be run as soon as this
is deployed.


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

---------

Co-authored-by: Gary Tou <gary@garytou.com>