[Referrals] Store referral program on login and permit custom styling and copy (#10953)
## Summary of the problem
Right now, we don't store `referral_program` on `logins`, which means we
need to store the referral as a `return_to`.

## Describe your changes
This refactors that code to be more in line with our convention of
storing login-based information on `login`.

It also allows `referral_programs` to provide custom text and a
background for the login page.