Fix admin create event flow to create contract for invitee (#10982)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->

Fixes https://github.com/hackclub/hcb/issues/10738

When creating a new event in the admin flow, invited users would not
have an associated contract created. This PR fixes that.

## Describe your changes

**Best reviewed by commits**

- extract a partial to ask for signee + cosigner + onboarding videos
from regular invite flow and share it with admin flow
- currently this has is_signee default to true since that was the
existing behavior for admins, we should revisit if that's what we want
for regular invites too. It's slightly awkward to set the default due to
the partial being shared now.

small QOL improvement to hide the cosigner field if the user is not
going to be a signee, as it's not relevant. This will still get sent to
the backend if a user filled it out but then flipped the `is_signee`
flag off, but the backend correctly ignores it.

We might want to fold in contract creation into the
OrganizerPositionInvite service somehow since they seem related.

---------

Co-authored-by: Albert Chae <albertchae@users.noreply.github.com>