[Announcements] Store custom elements as blocks in the database (#10992)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
The current way of storing custom elements (now "blocks") in JSON has
issues as it does not allow us to freeze the data or rendered HTML of
the blocks or update them individually.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Instead of storing blocks as JSON objects in the announcement content,
there is now a new `announcement_blocks` table that uses STI for each
type of block. Parameters for the blocks are stored in the `parameters`
column. This allows announcements to now display rendered block content
while editing as well as individually refreshing each block.

> [!WARNING]  
> A one time job `OneTimeJobs::MigrateAnnouncementBlocks` has been added
- **this should be run as soon as this PR is deployed, as announcements
using blocks will not render until this has been run.**

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

---------

Co-authored-by: Gary Tou <gary@garytou.com>