Add job and cron to delete old template draft announcements (#10970)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
A lot of database records are being created with #10945 merged, many of
which will not be used if the user does not click on the link to edit it
(which will convert it from a template draft into a draft).

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Every month, delete template draft announcements that are over a month
old.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->