Refactor mailer set_emails into an event method (#10997)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
The code to get a list of emails of all users in an event is replicated
across several mailers, and some don't take into account the team
contact email if it is set.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Added a method on `Event` that gets a list of all event users along with
the team contact email, and used it everywhere mailers need it.


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->