[Announcement] Small editor improvements (#11039)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
The refresh button was missing a tooltip and it was weird to have editor
padding but not be able to click on it to focus the editor.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Added a tooltip on the refresh button and added an action to focus the
editor when clicking the container.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->