Add time frames to home page graphs (#10530)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
The home page currently shows the text "ALL TIME" on the graphs, but
there's no way to change the time frame of the data shown.


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
This PR makes the current "ALL TIME" text on each home page graph a menu
that allows the user to select a valid time frame for the event. This is
passed to the breakdown engine which uses the provided time frame to
correctly calculate the sums.

The graphs, which were previously paired into turbo frames, are now
separate turbo frames to allow independently updating the time frame for
each graph.
<!-- If there are any visual changes, please attach images, videos, or
gifs. -->


![image](https://github.com/user-attachments/assets/c4b75790-8f88-43de-bdf7-143d7b3e5f54)

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Gary Tou <gary@garytou.com>