[Announcement] Add validation for unauthorized HCB code block (#11038)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
We're not currently doing any validations on when blocks are saved, and
leaving checking that any parameters are valid until we render the
block. Now that blocks are in the database, we can easily use
validations to not save any invalid data in the first place.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Added a validation to the HCB code block that makes sure that it is in
the same event as the announcement the block is a part of, and alerts it
to the user with a browser alert (this will be changed to something
nicer in a future PR). The existing checks in `render` are not removed
since invalid data could still be in the database or be added with a
Rails method that does not check validations.


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

---------

Co-authored-by: Gary Tou <gary@garytou.com>