[Announcement::Block::DonationSummary] Add end date and fallbacks to donation summary block (#11049)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
End dates for donation summaries are needed to properly support monthly
announcements and prevent it from updating on an accidental refresh.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Add the end date parameter to the donation summary block, and refactor a
bit to be better about default values for block parameters.


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->