[Announcement] Confirm publishing an announcement from the create form (#11150)
## Summary of the problem
<!-- Why are these changes being made? What problem does it solve? Link
any related issues to provide more details. -->
Currently we have `turbo_confirm` on publishing draft announcements, but
there's no confirmation when publishing an announcement straight from
the creation form!

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Adds a confirmation dialog to the creation form when publishing an
announcement for the first time.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->