[Announcement::Block::DonationGoal] Save event donation goal on block create (#11036)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
The donation goal block for announcements was not saving the ID of the
donation goal is was referencing, so if it was refreshed after a new
donation goal had been created, it would update to a different goal.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Saves the ID of the current donation goal in the parameters on save, and
references it if it exists.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

---------

Co-authored-by: David Cornu <davidc@hackclub.com>