[Admin] Separate admin user details into a turbo frame (#11002)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
Some user admin pages on production just don't load because of all of
the data that has to be fetched and rendered on that page, but only the
first couple admin tools are usually used.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Separates the slow-loading data on the user admin page into a
lazy-loaded turbo frame so that it doesn't hold back the other parts of
the page.
<!-- If there are any visual changes, please attach images, videos, or
gifs. -->