Fix `accordion` Stimulus controller usage (#11194)
## Summary of the problem

We're getting a couple of recurring errors in production:
1. The `content` target is sometimes missing
(https://appsignal.com/hack-club/sites/6596247683eb67648f30f807/exceptions/incidents/2025/samples/timestamp/2025-08-01T18:19:46Z)
    ```
    Missing target element "content" for "accordion" controller
    ```
2. We get console warnings about the target syntax
    ```
Please replace data-target="accordion.checkbox" with
data-accordion-target="checkbox". The data-target attribute is
deprecated and will be removed in a future version of Stimulus.
    ```
This was deprecated in
https://github.com/hotwired/stimulus/releases/tag/v2.0.0.

## Describe your changes

I audited the code base for places where we were using the `accordion`
controller and
- Updated `data-target` to `data-accordion-target`
- Removed references to the controller that didn't have a corresponding
`target` content (which is pretty essential to its functionality)