Hide unpublished monthly announcements (#11229)
## Summary of the problem
<!-- Why are these changes being made? What problem does it solve? Link
any related issues to provide more details. -->
Currently, after editing a monthly announcement, it will appear in the
announcement list. We don't want this to happen because it should only
be linked from the callout at the top of the page until it is published.


## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Remove announcements from the saved scope if their template type is
monthly and their published at is nil.


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->