Fix the amount and date filters on the transactions page (#11275)
Fixes https://github.com/hackclub/hcb/issues/11263

## Summary of the problem

### Context

- The transaction page uses a turbo frame to lazy load the ledger:
https://github.com/hackclub/hcb/blob/bf8c2791596a5d502be1b8d12046490445f3c110/app/views/events/transactions.html.erb#L144-L148
- The contents of the turbo frame are rendered by the `.../ledger`
endpoint (note that these contain the filters as well):
https://github.com/hackclub/hcb/blob/6e1deab06a36ee580aedc5b5abd8b4ed4a64b7b7/app/views/events/ledger.html.erb#L2-L3
- The filters are a mix of links
https://github.com/hackclub/hcb/blob/2abe39a5cde4869372dc8d4edbc302019753c6e5/app/views/events/_filter_menu.html.erb#L50-L52
and forms
https://github.com/hackclub/hcb/blob/2abe39a5cde4869372dc8d4edbc302019753c6e5/app/views/events/_filter_menu.html.erb#L57-L61
- When the user clicks on the filter icon, these filter options are
rendered inside of a menu
https://github.com/hackclub/hcb/blob/d84dbbac2d29d75362b7172df841cb4458e763bc/app/views/events/_filter_menu.html.erb#L1

### Where this went wrong

1. Our menu component supports a `data-menu-append-to-value` attribute
which determines where the DOM for the menu is rendered. This was
configured in this case but pointed to an ID that did not exist
(`turbo-frame#ledger`), which meant the menu was being appended outside
of the `<turbo-frame>` element.
2. This meant that when one of the filter forms was submitted, turbo
treated this as a full page visit to the `.../ledger` endpoint, and when
the response didn't include the full page structure it bailed and issued
a full page reload of that URL (likely because of the behaviour
described in
https://turbo.hotwired.dev/handbook/drive#reloading-when-assets-change)

## Describe your changes

- Fixes the `data-menu-append-to-value` attribute.
- Scopes the params passed to the ledger turbo frame to just
`request.query_parameters` instead of all of `params` as that led to
`action` and `controller` being set which caused havoc with all the
`link_to(upsert_query_params...` calls we had
- Uses explicit paths in all `link_to`s