Report `deprecation.rails` ActiveSupport::Notifications to AppSignal (#11289)
## Summary of the problem

We don't have much visibility into Rails deprecations in prod or a test
suite

## Describe your changes

HCB's setting for deprecations in production is notify

https://github.com/hackclub/hcb/blob/fd60bc13b0fe081f7e1da8cd4d0fa98ee2f21990/config/environments/production.rb#L115-L116

We can listen for these notifications and report them to AppSignal for
visibility. See
https://www.fastruby.io/blog/rails/upgrades/deprecation-warnings-rails-guide.html
for more details.

I think we should merge this soon so we can start collecting Rails 8
deprecations, then merge more of
https://github.com/hackclub/hcb/pull/10836 piecemeal to see if the
deprecations resolve.

To test locally, I changed the configuration to `:notify`, set a real
AppSignal credential and ran this in a rails console
```
ActiveRecord::Base.connection_pool.connection
```
This reported the following warning to AppSignal

```
DEPRECATION WARNING: ActiveRecord::ConnectionAdapters::ConnectionPool#connection is deprecated
and will be removed in Rails 8.0. Use #lease_connection instead.
```


https://appsignal.com/hack-club/sites/6596247683eb67648f30f807/exceptions/incidents/2122

---------

Co-authored-by: Albert Chae <albertchae@users.noreply.github.com>